feat: move code for layout using web worker to layout directory
